test: add test commands
Former-commit-id: 03cc67668121978d5b3dba05ac605cd7511f1920

+++ src/kernel/syscall.cpp
@@ -303,7 +303,7 @@ size_t sys_getcwd(Context *context) {
     }
     strcpy(ret, current_work_dir);
     // TODO: 如果用户使用虚拟地址此处会返回真实地址，这不合理
-    return (int) ((size_t) ret);
+    return (size_t) (ret);
 }
 
 size_t sys_dup(Context *context) {
